# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/apps/dubaiunitfinder/auth/register": {
      "post": {
        "description": "Register a new user for Dubai Unit Finder platform",
        "operationId": "DubaiunitfinderController_register",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for additional security (will be implemented later)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "password"
                ],
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "SecurePass123!"
                  },
                  "company": {
                    "type": "string",
                    "example": "Real Estate Co."
                  },
                  "referredBy": {
                    "type": "string",
                    "example": "ABC123XYZ456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "User registered successfully"
                    },
                    "userId": {
                      "type": "string",
                      "example": "uuid-here"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "User already exists or invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register new user",
        "tags": [
          "dubaiunitfinder"
        ]
      }
    },
    "/apps/dubaiunitfinder/auth/signin": {
      "post": {
        "description": "Sign in an existing user to Dubai Unit Finder platform",
        "operationId": "DubaiunitfinderController_signIn",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for additional security (will be implemented later)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "SecurePass123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sign in successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Sign in successful"
                    },
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sign in user",
        "tags": [
          "dubaiunitfinder"
        ]
      }
    },
    "/apps/dubaiunitfinder/subscription/{userId}": {
      "get": {
        "description": "Retrieve all subscriptions for a specific user",
        "operationId": "DubaiunitfinderController_getUserSubscriptions",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for additional security (will be implemented later)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscriptions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "subscriptions": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user subscriptions",
        "tags": [
          "dubaiunitfinder"
        ]
      }
    },
    "/apps/dubaiunitfinder/subscription": {
      "post": {
        "description": "Create a new subscription for a user",
        "operationId": "DubaiunitfinderController_createSubscription",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for additional security (will be implemented later)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "packageName",
                  "requestsInPackage",
                  "expiryDate"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "uuid-here"
                  },
                  "packageName": {
                    "type": "string",
                    "example": "premium"
                  },
                  "requestsInPackage": {
                    "type": "number",
                    "example": 1000
                  },
                  "expiryDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-12-31"
                  },
                  "price": {
                    "type": "number",
                    "example": 99.99
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Subscription created successfully"
                    },
                    "subscriptionId": {
                      "type": "string",
                      "example": "uuid-here"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid subscription data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new subscription",
        "tags": [
          "dubaiunitfinder"
        ]
      }
    },
    "/apps/dubaiunitfinder/history/{userId}": {
      "get": {
        "description": "Retrieve search history for a specific user",
        "operationId": "DubaiunitfinderController_getUserHistory",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for additional security (will be implemented later)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to return",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "x-api-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user history",
        "tags": [
          "dubaiunitfinder"
        ]
      }
    },
...
    "/apps/dubaiunitfinder/admin/users/{userId}": {
      "get": {
        "description": "Retrieve detailed information about a specific user",
        "operationId": "DubaiunitfinderController_getUserDetails",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for additional security (will be implemented later)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope or not admin"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user details (Admin)",
        "tags": [
          "dubaiunitfinder"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Gateway",
    "description": "API Gateway with OAuth2.0 Client Credentials Flow",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:9000/v2",
      "description": "Production server"
    },
    {
      "url": "http://localhost:9000/v2",
      "description": "Development server"
    },
    {
      "url": "http://localhost:9000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}